Add a --json global option flag
This will toggle the output to be JSON as opposed to the human readable
format.

It isn't a particularly pleasant fix, gobal options are removed from
the options list. Instead it must create a "command option" which will
be parsed for a second time with a default of 'true'

Eventually this should be more formally integrated into Commander