Websockify can be started for active sessions
If websockify and the screen grabber programs are not installed when the
desktop session is started they are not ran for that desktop session.
This commit introduces a new command `webify DESKTOP` that will start
websockify and the screen grabber for an active session.

To simplify the implementation, the screen grabber is only started if
the websockify process is about to be started.  It is therefore still
possible to have an active, web accessible desktop session for which the
screen grabber is not running and to have no way to fix this.