Fix detecting primary IP on Vagrant/VirtualBox
Vagrant running with VirtualBox uses 10.0.2.15 as the address of default
device for all machines.  This results in `flight-desktop` running on
openflight-vagrant-cluster being unable to distinguish between remote
and non-remote machines.

The solution to this is to avoid accepting 10.0.2.15 as the primary IP
for VirtualBox machines if there is a second address available.  This
fixes the issue of distinguishing remote and non-remote machines whilst
still supporting VNC connections from the Vagrant host and connections
via the Flight Desktop webapp.

The solution might not be suitable for uses of VirtualBox outside of
Vagrant or uses of Vagrant/VirtualBox outside of
openflight-vagrant-cluster.  However, I don't expect to OpenFlight to be
used on VagrantBox outside of our own development practices.