Post verify only when type unverfied -> verified
The post verification script is now only called if it transitions from
`unverfied` -> `verified`.  It may be desirable to also call a script
when a type transitions from `verified` -> `unverified`.  However, this
is not yet implemented and is not expected to be a common occurrence.

Along the way support for multiple post verification scripts has been
removed.  A single post verification script is currently sufficient and
supporting only one makes output handling and debugging simpler.