Allow configuring IP used for `DISPLAY`/`xrandr`
The IP address used in the `DISPLAY` environment variable used by the
`xrandr` command is now configurable.  If not set it will default to the
first private IP address for the session, falling back to the session's
primary IP address if it has no private IP addresses.  The order of the
session's IP addresses are determined by the `ip address show` command.

In addition to allowing the IP address to be configured, the default
behaviour is changed.  Previously, the session's primary IP address was
used, now the session's first private IP address is used.  It is assumed
that will be correct in most cases.